lab3 basic 1

+++ lab3_rework/include/uart.h
@@ -13,7 +13,7 @@ void  echo(char r);
 char* uart_gets(char* buffer);
 void  uart_puts(char *s);
 void  uart_puts_len(char *s, unsigned long len);
-void  uart_hex(unsigned int d);
+void  uart_hex(unsigned long long d);
 int   uart_int(int d);
 void  uart_printf();
